Watch namespaced contract configmaps
